make refactoring for images and utils
